Ref: extracting emission and making createEmission() method.
